Move prepare_invoke
