+++ starknet_py/net/account/account.py
@@ -1,7 +1,6 @@
 import dataclasses
 import re
-from collections import OrderedDict
-from typing import Dict, Iterable, List, Optional, Tuple, Union
+from typing import List, Optional, Union
 
 from starknet_py.common import create_compiled_contract
 from starknet_py.constants import FEE_CONTRACT_ADDRESS, QUERY_VERSION_BASE
@@ -30,15 +29,7 @@ from starknet_py.net.models.transaction import (
 from starknet_py.net.models.typed_data import TypedData
 from starknet_py.net.signer import BaseSigner
 from starknet_py.net.signer.stark_curve_signer import KeyPair, StarkCurveSigner
-from starknet_py.serialization.data_serializers.array_serializer import ArraySerializer
-from starknet_py.serialization.data_serializers.felt_serializer import FeltSerializer
-from starknet_py.serialization.data_serializers.payload_serializer import (
-    PayloadSerializer,
-)
-from starknet_py.serialization.data_serializers.struct_serializer import (
-    StructSerializer,
-)
-from starknet_py.utils.iterable import ensure_iterable
+from starknet_py.net.utils import prepare_invoke
 from starknet_py.utils.sync import add_sync_methods
 from starknet_py.utils.typed_data import TypedData as TypedDataDataclass
 
@@ -123,40 +114,6 @@ class Account(BaseAccount):
             )
 
         return max_fee
-
-    async def _prepare_invoke(
-        self,
-        calls: Calls,
-        max_fee: Optional[int] = None,
-        auto_estimate: bool = False,
-    ) -> Invoke:
-        """
-        Takes calls and creates Invoke from them.
-
-        :param calls: Single call or list of calls.
-        :param max_fee: Max amount of Wei to be paid when executing transaction.
-        :param auto_estimate: Use automatic fee estimation, not recommend as it may lead to high costs.
-        :return: Invoke created from the calls (without the signature).
-        """
-        nonce = await self.get_nonce()
-
-        call_descriptions, calldata = _merge_calls(ensure_iterable(calls))
-        wrapped_calldata = _execute_payload_serializer.serialize(
-            {"call_array": call_descriptions, "calldata": calldata}
-        )
-
-        transaction = Invoke(
-            calldata=wrapped_calldata,
-            signature=[],
-            max_fee=0,
-            version=self.supported_transaction_version,
-            nonce=nonce,
-            contract_address=self.address,
-        )
-
-        max_fee = await self._get_max_fee(transaction, max_fee, auto_estimate)
-
-        return _add_max_fee_to_transaction(transaction, max_fee)
 
     async def _verify_message_hash(self, msg_hash: int, signature: List[int]) -> bool:
         """
@@ -246,7 +203,16 @@ class Account(BaseAccount):
         max_fee: Optional[int] = None,
         auto_estimate: bool = False,
     ) -> Invoke:
-        execute_tx = await self._prepare_invoke(calls, max_fee, auto_estimate)
+        execute_tx = await prepare_invoke(
+            calls=calls,
+            nonce=await self.get_nonce(),
+            version=self.supported_transaction_version,
+            contract_address=self.address,
+        )
+
+        max_fee = await self._get_max_fee(execute_tx, max_fee, auto_estimate)
+        execute_tx = _add_max_fee_to_transaction(execute_tx, max_fee)
+
         signature = self.signer.sign_transaction(execute_tx)
         return _add_signature_to_transaction(execute_tx, signature)
 
@@ -430,42 +396,3 @@ def _add_max_fee_to_transaction(
     tx: TypeAccountTransaction, max_fee: int
 ) -> TypeAccountTransaction:
     return dataclasses.replace(tx, max_fee=max_fee)
-
-
-def _parse_call(call: Call, entire_calldata: List) -> Tuple[Dict, List]:
-    _data = {
-        "to": call.to_addr,
-        "selector": call.selector,
-        "data_offset": len(entire_calldata),
-        "data_len": len(call.calldata),
-    }
-    entire_calldata += call.calldata
-
-    return _data, entire_calldata
-
-
-def _merge_calls(calls: Iterable[Call]) -> Tuple[List[Dict], List[int]]:
-    call_descriptions = []
-    entire_calldata = []
-    for call in calls:
-        data, entire_calldata = _parse_call(call, entire_calldata)
-        call_descriptions.append(data)
-
-    return call_descriptions, entire_calldata
-
-
-_felt_serializer = FeltSerializer()
-_call_description = StructSerializer(
-    OrderedDict(
-        to=_felt_serializer,
-        selector=_felt_serializer,
-        data_offset=_felt_serializer,
-        data_len=_felt_serializer,
-    )
-)
-_execute_payload_serializer = PayloadSerializer(
-    OrderedDict(
-        call_array=ArraySerializer(_call_description),
-        calldata=ArraySerializer(_felt_serializer),
-    )
-)
